#41 Rename vite config
